Added file fixation based on reviewer comments
